Twitter hackery: subcontexts, smartnarrow by responsible party.
This is for people like me who follow people on Twitter who do a lot of
retweeting.  I'd like to be able to see the person responsible for
putting the thing in front of my face, not just the original author; and
I'd like to smartnarrow like that, for similar reasons.  (TODO: put a
more verbose rationale here.)

Things that would need to be done to get this upstream: make it all
conditional, because there are probably people who like the current
behavior and dislike what I've done.  (Assuming anybody else actually
uses barnowl twitter in the first place.)